# tx-op validation without using spec