--- a/crux-core/src/crux/api.clj
+++ b/crux-core/src/crux/api.clj
@@ -17,58 +17,13 @@
 (s/def :crux.db.fn/body (s/cat :fn #{'fn}
                                :args (s/coll-of symbol? :kind vector? :min-count 1)
                                :body (s/* any?)))
-(s/def ::doc (s/and (s/map-of keyword? any?)
-                    (s/keys :req [:crux.db/id] :opt [:crux.db.fn/body :crux.db.fn/args])))
 
-(def ^:private date? (partial instance? Date))
-
-(defmulti tx-op first)
-
-(defmethod tx-op :crux.tx/put [_]
-  (s/cat :op #{:crux.tx/put}
-         :doc ::doc
-         :start-valid-time (s/? date?)
-         :end-valid-time (s/? date?)))
-
-(defmethod tx-op :crux.tx/delete [_]
-  (s/cat :op #{:crux.tx/delete}
-         :id :crux.db/id
-         :start-valid-time (s/? date?)
-         :end-valid-time (s/? date?)))
-
-(defmethod tx-op :crux.tx/cas [_]
-  (s/cat :op #{:crux.tx/cas}
-         :old-doc (s/nilable ::doc)
-         :new-doc ::doc
-         :at-valid-time (s/? date?)))
-
-(defmethod tx-op :crux.tx/match [_]
-  (s/cat :op #{:crux.tx/match}
-         :id :crux.db/id
-         :doc (s/nilable ::doc)
-         :at-valid-time (s/? date?)))
-
-(defmethod tx-op :crux.tx/evict [_]
-  (s/cat :op #{:crux.tx/evict}
-         :id :crux.db/id))
-
-(s/def ::args-doc (s/and ::doc (s/keys :req [:crux.db.fn/args])))
-
-(defmethod tx-op :crux.tx/fn [_]
-  (s/cat :op #{:crux.tx/fn}
-         :id :crux.db/id
-         :args-doc (s/? ::args-doc)))
-
-(s/def ::tx-op (s/multi-spec tx-op first))
-(s/def ::tx-ops (s/coll-of ::tx-op :kind vector?))
-
-(defn- conform-tx-ops  [tx-ops]
+(defn- conform-tx-ops [tx-ops]
   (->> tx-ops
        (mapv
         (fn [tx-op]
-          (map
-           #(if (instance? java.util.Map %) (into {} %) %)
-           tx-op)))
+          (map #(if (instance? java.util.Map %) (into {} %) %)
+               tx-op)))
        (mapv vec)))
 
 (defprotocol PCruxNode